requirements.txt
python-multipart==0.0.6
requests==2.31.0
httptools==0.5.0
python-dotenv==1.0.0
redis==4.6.0
loguru==0.7.2
pydantic==2.4.2
cryptography==41.0.4
celery==5.3.4
SQLAlchemy==1.4.5
psycopg2-binary==2.9.9
netifaces
minio==7.2.0
Pillow==10.2.0

tqdm==4.66.1
numpy==1.26.3
torch==2.1.2
peft==0.8.1
imageio[ffmpeg]

accelerate==0.25.0
transformers==4.38.1
diffusers[torch]==0.25.0
ftfy==6.1.1
opencv-python==4.7.0.68
einops==0.6.1
pytorch-lightning==1.9.0
tensorboard==2.10.1
safetensors>=0.4.1
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.20.1
xformers
open-clip-torch==2.20.0

av
flashy
hydra-core
hydra_colorlog
julius
num2words
sentencepiece
spacy
torchaudio
demucs
librosa
gradio
torchmetrics
encodec
protobuf
audiocraft
peft
datasets
trl